Render the ghost in the slot of the block it becomes
The ghost joins the client IDs the list maps over instead of a sibling
slot, and the materialise listeners live inside the block provider, so
the element before and after insertion is the same component at the
same key: nothing remounts and focus survives keyboard entry. The
default block resolves from the block list settings, and the import
headers follow the new dependency rules.
